+++ cmake/build.cmake
@@ -95,6 +95,7 @@ add_executable(${TARGET}
   src/flutter/shell/platform/linux_embedded/plugin/text_input_plugin.cc
   src/flutter/shell/platform/linux_embedded/plugin/platform_plugin.cc
   src/flutter/shell/platform/linux_embedded/plugin/mouse_cursor_plugin.cc
+  src/flutter/shell/platform/linux_embedded/surface/context_egl.cc
   src/flutter/shell/platform/linux_embedded/surface/egl_utils.cc
   ${DISPLAY_BACKEND_SRC}
   ${WAYLAND_PROTOCOL_SRC}
